New CDN publish workflow (#30009)
* LET'S SEE IF THIS WORKS

* I forgot the chmod +x

* I forgot the shebang
