+++ .github/workflows/publish.yml
@@ -41,31 +41,22 @@ jobs:
     - name: Package client
       run: dotnet run --project Content.Packaging client --no-wipe-release
 
-    - name: Update Build Info
-      run: Tools/gen_build_info.py
-
-    - name: Shuffle files around
-      run: |
-        mkdir "release/${{ github.sha }}"
-        mv release/*.zip "release/${{ github.sha }}"
-
-    - name: Upload files to centcomm
-      uses: appleboy/scp-action@master
+    - name: Upload build artifact
+      id: artifact-upload-step
+      uses: actions/upload-artifact@v4
       with:
-        host: centcomm.spacestation14.io
-        username: wizards-build-push
-        key: ${{ secrets.CENTCOMM_WIZARDS_BUILDS_PUSH_KEY }}
-        source: "release/${{ github.sha }}"
-        target: "/home/wizards-build-push/builds_dir/builds/"
-        strip_components: 1
+        name: build
+        path: release/*.zip
+        compression-level: 0
+        retention-days: 0
 
-    - name: Update manifest JSON
-      uses: appleboy/ssh-action@master
-      with:
-        host: centcomm.spacestation14.io
-        username: wizards-build-push
-        key: ${{ secrets.CENTCOMM_WIZARDS_BUILDS_PUSH_KEY }}
-        script: /home/wizards-build-push/push.ps1 ${{ github.sha }}
+    - name: Publish version
+      run: Tools/publish_github_artifact.py
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        PUBLISH_TOKEN: ${{ secrets.PUBLISH_TOKEN }}
+        ARTIFACT_ID: ${{ steps.artifact-upload-step.outputs.artifact-id }}
+        GITHUB_REPOSITORY: ${{ vars.GITHUB_REPOSITORY }}
 
     - name: Publish changelog (Discord)
       run: Tools/actions_changelogs_since_last_run.py
@@ -77,3 +68,8 @@ jobs:
       run: Tools/actions_changelog_rss.py
       env:
         CHANGELOG_RSS_KEY: ${{ secrets.CHANGELOG_RSS_KEY }}
+
+    - uses: geekyeggo/delete-artifact@v5
+      if: always()
+      with:
+        name: build
